+++ lib/money.rb
@@ -30,5 +30,7 @@ require 'money/core_extensions'
 # FIXME: Consider usage of specific major version (Rails::VERSION::MAJOR == 3)
 if defined?(Rails)
   require 'money/monetizable/monetize'
-  require 'money/monetizable/railtie'
+  
+	require 'money/monetizable/engine' if defined?(::Rails::Engine)
+  require 'money/monetizable/railtie' if defined?(::Rails::Railtie) && !defined?(::Rails::Engine)
 end