+++ lib/money/monetizable/railtie.rb
@@ -1,5 +1,0 @@
-module Monetizable
-  class Railtie < Rails::Railtie
-    ActiveRecord::Base.send :include, Monetizable
-  end
-end