+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 5.1.0
+
+- Engine now supports multiple ORMs and adds monetizable to each Model base module/class if present
+- Railtie converted to Engine
+- merged with alup monetizable branch with Railtie and Monetizable
+
 ## 5.0.0
 - Money::Currency::TABLE removed. Use Money::Currency.register to add
   additional currencies (GH-143)